+++ check_scannability.py
@@ -19,6 +19,7 @@ from datetime import datetime
 from doc_utils.scannability import check_scannability
 from doc_utils.version_check import check_version_on_startup
 from doc_utils.file_utils import collect_files, parse_exclude_list_file
+from doc_utils.version import __version__
 
 from doc_utils.spinner import Spinner
 BASE_SENTENCE_WORD_LIMIT = 22
@@ -43,6 +44,7 @@ def main():
     parser.add_argument('--exclude-dir', action='append', default=[], help='Directory to exclude (can be used multiple times).')
     parser.add_argument('--exclude-file', action='append', default=[], help='File to exclude (can be used multiple times).')
     parser.add_argument('--exclude-list', type=str, help='Path to a file containing directories or files to exclude, one per line.')
+    parser.add_argument('--version', action='version', version=f'%(prog)s {__version__}')
     # Do not add -h/--help to argparse, handled manually above
     args = parser.parse_args()